# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: MiBee Steward API
  description: |
    Device management & monitoring system (CMDB-lite for network/IoT assets):
    auto-discovery, identity inference, registry, heartbeat, synthetic
    probing, distributed agents.

    Auth for browser/user routes: JWT, cookie-first with Bearer header
    fallback (`Authorization: Bearer <token>`). Machine routes (the
    /api/v1/agents/{report,probe-report,commands,commands/{id}/...} family)
    use per-agent bearer tokens minted by an admin.
  version: 0.6.0
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /devices:
    get:
      tags: [Devices]
      summary: List devices (filters status/type/search/sort/order/network_id/created_from/created_to)
      parameters:
        - {$ref: "#/components/parameters/limit"}
        - {$ref: "#/components/parameters/offset"}
        - {name: status, in: query, schema: {type: string}}
        - {name: type, in: query, schema: {type: string}}
        - {name: search, in: query, schema: {type: string}}
        - {name: sort, in: query, schema: {type: string}}
        - {name: order, in: query, schema: {type: string, enum: [asc, desc]}}
        - {name: network_id, in: query, schema: {type: integer}}
      responses:
        "200":
          description: Device list envelope
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DeviceList"}
    post:
      tags: [Devices]
      summary: Create a device (manual entry)
      responses:
        "201":
          description: Created device
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Device"}
...
  /devices/{id}:
    get:
      tags: [Devices]
      summary: One device (out-of-scope = 404)
      parameters: [{$ref: "#/components/parameters/id"}]
      responses:
        "200":
          description: Device
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Device"}
        "404": {$ref: "#/components/responses/Error"}
    put:
      tags: [Devices]
      summary: Update editable device fields
      parameters: [{$ref: "#/components/parameters/id"}]
      responses:
        "200":
          description: Updated device
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Device"}
    delete:
      tags: [Devices]
      summary: Delete a device (cascades satellites)
      parameters: [{$ref: "#/components/parameters/id"}]
      responses:
        "204": {description: Deleted}
...
  /documents/{id}/devices:
    get:
      tags: [Documents]
      summary: Devices linked to a document ({devices,total})
      parameters: [{$ref: "#/components/parameters/id"}]
      responses:
        "200": {description: Linked-device envelope (complete list)}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: JWT from POST /auth/login (Bearer fallback when cookies are unavailable)
    cookieAuth:
      type: apiKey
      in: cookie
      name: mibee_token
      description: JWT session cookie (preferred transport)
    agentToken:
      type: http
      scheme: bearer
      description: Per-agent token minted via POST /agents/tokens (machine channel)
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    limit:
      name: limit
      in: query
      description: Page size (absent/0 → endpoint default; malformed or negative → 400; over-max clamps)
      schema: {type: integer, format: int64, minimum: 0}
    offset:
      name: offset
      in: query
      description: Pagination offset (negative → 400)
      schema: {type: integer, format: int64, minimum: 0}
  responses:
...
    Error:
      description: Error envelope
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable message (single-field error model)
    PageMeta:
      type: object
      description: Echo of the effective pagination (absent on complete lists)
      properties:
        total: {type: integer, format: int64, description: Filtered count}
        limit: {type: integer, format: int64}
        offset: {type: integer, format: int64}
...
    SNMPDiscovery:
      type: object
      description: Parsed SNMP sysObject query (mirrors domain.SNMPDiscovery)
      properties:
        sys_descr: {type: string}
        sys_object_id: {type: string}
        sys_name: {type: string}
        sys_location: {type: string}
        sys_contact: {type: string}
        sys_services: {type: integer}
    OpenPortEntry:
      type: object
      required: [port]
      properties:
        port: {type: integer}
        service: {type: string}
    ServiceEntry:
      type: object
      required: [port, name]
      properties:
        port: {type: integer}
        name: {type: string}
        protocol: {type: string}
        version: {type: string}
    PrometheusInfo:
      type: object
      properties:
        url: {type: string}
        node_exporter_url: {type: string}
        labels: {type: object, additionalProperties: {type: string}}
    ScanAttributes:
      type: object
      description: >-
        Engine-written discovery document (mirrors domain.ScanAttributes; stored
        as JSON in devices.scan_attributes). All fields fill progressively,
        treat every field as optional. Engine-owned: user edits go to
        user_attributes.
      additionalProperties: true
      properties:
        vendor: {type: string, description: OUI lookup or SNMP/HTTP-derived vendor}
        mac: {type: string, description: "normalized lowercase aa:bb:cc:.."}
        hostname: {type: string}
        mac_is_locally_administered: {type: boolean, description: U/L bit set, neutral observability flag, no identity effect}
        mac_is_multicast: {type: boolean}
        oui_prefix: {type: string, description: "IEEE block via longest-prefix (6/7/9 hex)"}
        oui_vendor: {type: string, description: NIC silicon vendor, distinct from self-declared vendor}
        os: {type: string}
        os_version: {type: string}
        kernel_version: {type: string}
        firmware_version: {type: string}
        cpu_count: {type: integer}
        cpu_model: {type: string}
        memory_total_bytes: {type: integer, format: int64}
        uptime_seconds: {type: integer, format: int64}
        ttl: {type: integer}
        last_scan_rtt_ms: {type: integer, format: int64}
        scan_source: {type: string}
        last_scanned_at: {type: string, description: RFC3339; empty when never scanned}
        inferred_type: {type: string}
        inferred_type_source: {type: string, enum: ["", protocol, heuristic], description: "protocol = evidence-backed (trustworthy); heuristic = hostname/brand guess (spoofable)"}
        inferred_description: {type: string}
        snmp: {$ref: "#/components/schemas/SNMPDiscovery"}
        open_ports: {type: array, items: {$ref: "#/components/schemas/OpenPortEntry"}}
        detected_services: {type: array, items: {$ref: "#/components/schemas/ServiceEntry"}}
        prometheus: {$ref: "#/components/schemas/PrometheusInfo"}
        extras: {type: object, additionalProperties: {type: string}, description: Namespaced probe overflow (mdns.*, ssdp.*, ...)}
    Device:
      type: object
      required: [id, name, type, brand, model, location, purpose, description, status, ip_address, mac_address, serial_number, purchase_date, warranty_expiry, tags, created_at, updated_at, scan_source, prometheus_labels, open_ports, detected_services, prometheus_url, node_exporter_url, last_scan_rtt_ms, scan_attributes, user_attributes]
      properties:
        id: {type: integer, format: int64}
        name: {type: string}
        type: {type: string, enum: [pc, embedded, iot, server, switch, router, firewall, nas, camera, phone, printer, other]}
        brand: {type: string}
        model: {type: string}
        location: {type: string}
        purpose: {type: string}
        description: {type: string}
        status: {type: string, enum: [online, offline, unknown]}
        ip_address: {type: string}
        mac_address: {type: string}
        serial_number: {type: string}
        purchase_date: {type: string}
        warranty_expiry: {type: string}
        tags: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        scan_source: {type: string, description: "scanner_v2 | manual | passive:*"}
        network_id: {type: integer, format: int64, nullable: true, description: Logical network the device was discovered on (absent for legacy rows)}
        network_name: {type: string}
        prometheus_labels: {type: string}
        last_scanned_at: {type: string, format: date-time, nullable: true}
        last_scan_task_id: {type: integer, format: int64, nullable: true}
        last_seen: {type: string, format: date-time, nullable: true, description: Scan-derived last observed online}
        last_online_at: {type: string, format: date-time, nullable: true, description: Authoritative last-alive from the verdict series (detail rows only)}
        offline_since: {type: string, format: date-time, nullable: true, description: Retention clock start when offline}
        open_ports: {type: string, description: "Legacy JSON-string column; parse for OpenPortEntry items. UI prefers scan_attributes.open_ports"}
        detected_services: {type: string, description: "Legacy JSON-string column; parse for ServiceEntry items. UI prefers scan_attributes.detected_services"}
        prometheus_url: {type: string}
        node_exporter_url: {type: string}
        last_scan_rtt_ms: {type: integer, format: int64}
        scan_attributes: {$ref: "#/components/schemas/ScanAttributes"}
        user_attributes: {type: object, additionalProperties: {type: string}}
    DeviceList:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            devices:
              type: array
              items: {$ref: "#/components/schemas/Device"}
...